fix: fix package declarations
Fixes package declarations and imports of classes that had been moved
but had not changed their package declarations. Also moves test classes
that need to be in the same package as the class that they are testing.
