+++ src/main/java/liquibase/ext/spanner/change/MergeColumnsChangeSpanner.java
@@ -11,7 +11,7 @@
  * express or implied. See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package liquibase.ext.spanner;
+package liquibase.ext.spanner.change;
 
 import java.util.ArrayList;
 import java.util.Arrays;
