+++ src/main/java/liquibase/ext/spanner/datatype/CharTypeSpanner.java
@@ -11,11 +11,12 @@
  * express or implied. See the License for the specific language governing permissions and
  * limitations under the License. package liquibase.ext.spanner;
  */
-package liquibase.ext.spanner;
+package liquibase.ext.spanner.datatype;
 
 import liquibase.database.Database;
 import liquibase.datatype.DatabaseDataType;
 import liquibase.datatype.core.CharType;
+import liquibase.ext.spanner.CloudSpanner;
 
 /** CHAR(n) is translated to STRING(n). */
 public class CharTypeSpanner extends CharType {
