+++ src/main/java/liquibase/ext/spanner/datatype/DoubleTypeSpanner.java
@@ -1,8 +1,9 @@
-package liquibase.ext.spanner;
+package liquibase.ext.spanner.datatype;
 
 import liquibase.database.Database;
 import liquibase.datatype.DatabaseDataType;
 import liquibase.datatype.core.DoubleType;
+import liquibase.ext.spanner.CloudSpanner;
 
 public class DoubleTypeSpanner extends DoubleType {
   private static final DatabaseDataType FLOAT64 = new DatabaseDataType("FLOAT64");
