+++ src/main/java/liquibase/ext/spanner/datatype/UuidTypeSpanner.java
@@ -11,11 +11,12 @@
  * express or implied. See the License for the specific language governing permissions and
  * limitations under the License. package liquibase.ext.spanner;
  */
-package liquibase.ext.spanner;
+package liquibase.ext.spanner.datatype;
 
 import liquibase.database.Database;
 import liquibase.datatype.DatabaseDataType;
 import liquibase.datatype.core.UUIDType;
+import liquibase.ext.spanner.CloudSpanner;
 
 /** UUID is translated to STRING(36) as Cloud Spanner does not have a built-in type for UUID's. */
 public class UuidTypeSpanner extends UUIDType {
