+++ src/main/java/liquibase/ext/spanner/datatype/XmlTypeSpanner.java
@@ -11,11 +11,12 @@
  * express or implied. See the License for the specific language governing permissions and
  * limitations under the License. package liquibase.ext.spanner;
  */
-package liquibase.ext.spanner;
+package liquibase.ext.spanner.datatype;
 
 import liquibase.database.Database;
 import liquibase.datatype.DatabaseDataType;
 import liquibase.datatype.core.XMLType;
+import liquibase.ext.spanner.CloudSpanner;
 
 /** XML is translated to STRING(MAX) as Cloud Spanner does not have a built-in type for XML. */
 public class XmlTypeSpanner extends XMLType {
