+++ src/main/java/liquibase/ext/spanner/sqlgenerator/AddUniqueConstraintGeneratorSpanner.java
@@ -11,10 +11,11 @@
  * express or implied. See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package liquibase.ext.spanner;
+package liquibase.ext.spanner.sqlgenerator;
 
 import liquibase.database.Database;
 import liquibase.exception.ValidationErrors;
+import liquibase.ext.spanner.CloudSpanner;
 import liquibase.sqlgenerator.SqlGeneratorChain;
 import liquibase.sqlgenerator.core.AddUniqueConstraintGenerator;
 import liquibase.statement.core.AddUniqueConstraintStatement;
