+++ src/main/java/liquibase/ext/spanner/sqlgenerator/InsertOrUpdateGeneratorSpanner.java
@@ -14,11 +14,12 @@
  * or implied. See the License for the specific language governing permissions and limitations under
  * the License.
  */
-package liquibase.ext.spanner;
+package liquibase.ext.spanner.sqlgenerator;
 
 import java.util.ArrayList;
 import liquibase.database.Database;
 import liquibase.exception.LiquibaseException;
+import liquibase.ext.spanner.CloudSpanner;
 import liquibase.sql.Sql;
 import liquibase.sql.UnparsedSql;
 import liquibase.sqlgenerator.SqlGeneratorChain;
