+++ src/main/java/liquibase/ext/spanner/sqlgenerator/SetNullableGeneratorSpanner.java
@@ -14,10 +14,11 @@
  * or implied. See the License for the specific language governing permissions and limitations under
  * the License.
  */
-package liquibase.ext.spanner;
+package liquibase.ext.spanner.sqlgenerator;
 
 import liquibase.database.Database;
 import liquibase.datatype.DataTypeFactory;
+import liquibase.ext.spanner.CloudSpanner;
 import liquibase.sql.Sql;
 import liquibase.sql.UnparsedSql;
 import liquibase.sqlgenerator.SqlGenerator;
