+++ src/test/java/liquibase/ext/spanner/datatype/VarcharTypeSpannerTest.java
@@ -1,9 +1,10 @@
-package liquibase.ext.spanner;
+package liquibase.ext.spanner.datatype;
 
 import static com.google.common.truth.Truth.assertThat;
 import static org.mockito.Mockito.mock;
 
 import liquibase.database.core.PostgresDatabase;
+import liquibase.ext.spanner.CloudSpanner;
 import org.junit.jupiter.api.Test;
 
 public class VarcharTypeSpannerTest {
