+++ src/test/java/liquibase/ext/spanner/sqlgenerator/CreateDropProcedureTest.java
@@ -11,7 +11,7 @@
  * express or implied. See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package liquibase.ext.spanner;
+package liquibase.ext.spanner.sqlgenerator;
 
 import static com.google.common.truth.Truth.assertThat;
 import static org.junit.Assert.fail;
@@ -19,14 +19,16 @@ import static org.junit.Assert.fail;
 import java.sql.Connection;
 import liquibase.Contexts;
 import liquibase.Liquibase;
+import liquibase.exception.LiquibaseException;
 import liquibase.exception.ValidationFailedException;
+import liquibase.ext.spanner.AbstractMockServerTest;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.parallel.Execution;
 import org.junit.jupiter.api.parallel.ExecutionMode;
 
 @Execution(ExecutionMode.SAME_THREAD)
-public class RenameColumnTest extends AbstractMockServerTest {
+public class CreateDropProcedureTest extends AbstractMockServerTest {
 
   @BeforeEach
   void resetServer() {
@@ -35,15 +37,29 @@ public class RenameColumnTest extends AbstractMockServerTest {
   }
 
   @Test
-  void testAddPrimaryKeySingersFromYaml() throws Exception {
-    for (String file : new String[] {"rename-column-singers.spanner.yaml"}) {
+  void testCreateProcedureFromYaml() throws Exception {
+    for (String file : new String[] {"create-procedure.spanner.yaml"}) {
+      try (Connection con = createConnection();
+          Liquibase liquibase = getLiquibase(con, file)) {
+        liquibase.update(new Contexts("test"));
+        fail("missing expected validation exception");
+      } catch (LiquibaseException e) {
+        assertThat(e.getMessage()).contains(CreateProcedureGeneratorSpanner.CREATE_PROCEDURE_VALIDATION_ERROR);
+      }
+    }
+    assertThat(mockAdmin.getRequests()).isEmpty();
+  }
+
+  @Test
+  void testDropProcedureFromYaml() throws Exception {
+    for (String file : new String[] {"drop-procedure.spanner.yaml"}) {
       try (Connection con = createConnection();
           Liquibase liquibase = getLiquibase(con, file)) {
         liquibase.update(new Contexts("test"));
         fail("missing expected validation exception");
       } catch (ValidationFailedException e) {
         assertThat(e.getMessage())
-            .contains(RenameColumnGeneratorSpanner.RENAME_COLUMN_VALIDATION_ERROR);
+            .contains(DropProcedureGeneratorSpanner.DROP_PROCEDURE_VALIDATION_ERROR);
       }
     }
     assertThat(mockAdmin.getRequests()).isEmpty();
