+++ src/test/java/liquibase/ext/spanner/sqlgenerator/RenameTableTest.java
@@ -11,7 +11,7 @@
  * express or implied. See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package liquibase.ext.spanner;
+package liquibase.ext.spanner.sqlgenerator;
 
 import static com.google.common.truth.Truth.assertThat;
 import static org.junit.Assert.fail;
@@ -20,13 +20,15 @@ import java.sql.Connection;
 import liquibase.Contexts;
 import liquibase.Liquibase;
 import liquibase.exception.ValidationFailedException;
+import liquibase.ext.spanner.AbstractMockServerTest;
+import liquibase.ext.spanner.sqlgenerator.RenameTableGeneratorSpanner;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.parallel.Execution;
 import org.junit.jupiter.api.parallel.ExecutionMode;
 
 @Execution(ExecutionMode.SAME_THREAD)
-public class AddDropDefaultValueTest extends AbstractMockServerTest {
+public class RenameTableTest extends AbstractMockServerTest {
 
   @BeforeEach
   void resetServer() {
@@ -35,30 +37,15 @@ public class AddDropDefaultValueTest extends AbstractMockServerTest {
   }
 
   @Test
-  void testAddDefaultValueSingersFromYaml() throws Exception {
-    for (String file : new String[] {"add-default-value-singers.spanner.yaml"}) {
+  void testRenameTableSingersFromYaml() throws Exception {
+    for (String file : new String[] {"rename-table-singers.spanner.yaml"}) {
       try (Connection con = createConnection();
           Liquibase liquibase = getLiquibase(con, file)) {
         liquibase.update(new Contexts("test"));
         fail("missing expected validation exception");
       } catch (ValidationFailedException e) {
         assertThat(e.getMessage())
-            .contains(AddDefaultValueGeneratorSpanner.ADD_DEFAULT_VALUE_VALIDATION_ERROR);
-      }
-    }
-    assertThat(mockAdmin.getRequests()).isEmpty();
-  }
-
-  @Test
-  void testDropDefaultValueSingersFromYaml() throws Exception {
-    for (String file : new String[] {"drop-default-value-singers.spanner.yaml"}) {
-      try (Connection con = createConnection();
-          Liquibase liquibase = getLiquibase(con, file)) {
-        liquibase.update(new Contexts("test"));
-        fail("missing expected validation exception");
-      } catch (ValidationFailedException e) {
-        assertThat(e.getMessage())
-            .contains(DropDefaultValueGeneratorSpanner.DROP_DEFAULT_VALUE_VALIDATION_ERROR);
+            .contains(RenameTableGeneratorSpanner.RENAME_TABLE_VALIDATION_ERROR);
       }
     }
     assertThat(mockAdmin.getRequests()).isEmpty();
